+++ backend/src/assets/assets.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { AssetsController } from './assets.controller';
 import { AssetsService } from './assets.service';
 import { InterventionsModule } from '../interventions/interventions.module';
+import { WarrantiesModule } from '../warranties/warranties.module';
 
 @Module({
-  imports: [InterventionsModule],
+  imports: [InterventionsModule, WarrantiesModule],
   controllers: [AssetsController],
   providers: [AssetsService],
   exports: [AssetsService],